MRF penalty from a SpatialPoylgonsDataFrame

MRF penalty from a SpatialPoylgonsDataFrame

# S3 method for SpatialPolygonsDataFrame
mrf_penalty(object, node_labels = NULL, buffer = NULL, add_delta = FALSE, ...)

Arguments

object

an R object to create the MRF penalty from.

node_labels

character; a vector of alternative labels for the levels of the factor.

buffer

numeric; buffer distance for all or for individual elements of the geometry. See argument dist in sf::st_buffer for details.

add_delta

numeric or logical; either the numeric value to add to the diagonal of the MRF penalty matrix, or a logical value indicating if such an adjustment should be made. The default is to not alter the diagonal of the penalty matrix.

...

arguments passed to other methods.