First-order random walk MRF penalty from a numeric vector

First-order random walk MRF penalty from a numeric vector

# S3 method for numeric
mrf_penalty(
  object,
  type = c("linear", "cyclic"),
  node_labels = NULL,
  add_delta = FALSE,
  end_points = NULL,
  ...
)

Arguments

object

an R object to create the MRF penalty from.

type

character; one of "linear" or "cyclic" indicating if the observations form a cyclic series or not.

node_labels

character; a vector of alternative labels for the levels of the factor.

add_delta

numeric or logical; either the numeric value to add to the diagonal of the MRF penalty matrix, or a logical value indicating if such an adjustment should be made. The default is to not alter the diagonal of the penalty matrix.

end_points

numeric; an optional vector of length 2 providing the end points of the period of cycle.

...

arguments passed to other methods.

Examples

# linear
p <- mrf_penalty(1:10)
as.matrix(p)
#>     1  2  3  4  5  6  7  8  9 10
#> 1   1 -1  0  0  0  0  0  0  0  0
#> 2  -1  2 -1  0  0  0  0  0  0  0
#> 3   0 -1  2 -1  0  0  0  0  0  0
#> 4   0  0 -1  2 -1  0  0  0  0  0
#> 5   0  0  0 -1  2 -1  0  0  0  0
#> 6   0  0  0  0 -1  2 -1  0  0  0
#> 7   0  0  0  0  0 -1  2 -1  0  0
#> 8   0  0  0  0  0  0 -1  2 -1  0
#> 9   0  0  0  0  0  0  0 -1  2 -1
#> 10  0  0  0  0  0  0  0  0 -1  1

# cyclic
p <- mrf_penalty(1:10, type = "cyclic")
as.matrix(p)
#>          1  2  3  4  5  6  7  8  9      10
#> 1   500001 -1  0  0  0  0  0  0  0 -500000
#> 2       -1  2 -1  0  0  0  0  0  0       0
#> 3        0 -1  2 -1  0  0  0  0  0       0
#> 4        0  0 -1  2 -1  0  0  0  0       0
#> 5        0  0  0 -1  2 -1  0  0  0       0
#> 6        0  0  0  0 -1  2 -1  0  0       0
#> 7        0  0  0  0  0 -1  2 -1  0       0
#> 8        0  0  0  0  0  0 -1  2 -1       0
#> 9        0  0  0  0  0  0  0 -1  2      -1
#> 10 -500000  0  0  0  0  0  0  0 -1  500001

# cyclic with user end points
p <- mrf_penalty(1:10, type = "cyclic", end_points = c(0,11))
as.matrix(p)
#>       1  2  3  4  5  6  7  8  9   10
#> 1   1.5 -1  0  0  0  0  0  0  0 -0.5
#> 2  -1.0  2 -1  0  0  0  0  0  0  0.0
#> 3   0.0 -1  2 -1  0  0  0  0  0  0.0
#> 4   0.0  0 -1  2 -1  0  0  0  0  0.0
#> 5   0.0  0  0 -1  2 -1  0  0  0  0.0
#> 6   0.0  0  0  0 -1  2 -1  0  0  0.0
#> 7   0.0  0  0  0  0 -1  2 -1  0  0.0
#> 8   0.0  0  0  0  0  0 -1  2 -1  0.0
#> 9   0.0  0  0  0  0  0  0 -1  2 -1.0
#> 10 -0.5  0  0  0  0  0  0  0 -1  1.5